requirements.txt
aiohttp==2.0.7
async-timeout==1.2.1
future==0.16.0
python-telegram-bot==5.3.1
telepot==11.0
virtualenv==15.1.0

